LSM6DSO32X imu_read_sample: real accel/gyro read (silicon-verified)
Graduates the bench-verified bring-up sequence into the Rev 6 IMU driver:
imu_init does a WHO_AM_I check (0x6C) then configures CTRL3_C=0x44
(BDU+IF_INC), CTRL1_XL=0x40 (104 Hz, ±4 g), CTRL2_G=0x44 (104 Hz,
500 dps). imu_read_sample checks STATUS data-ready then burst-reads the
12-byte gyro+accel block at 0x22 (IF_INC auto-increment) and converts at
the datasheet sensitivities (0.122 mg/LSB, 17.5 mdps/LSB).

±4 g suits the Rev 6 interpretable features (gravity-derived posture,
gait, activity RMS); true impacts are the Rev 7 ±320 g channel's job.
Widening to ±16 g is a one-line CTRL1_XL change if launch data wants it.

Also widens imu_sample_t accel/gyro fields from int16 to int32: at
500 dps full-scale a gyro reading reaches ~573000 mdps, far past int16's
±32767 (the old type would have wrapped). Now matches the SINT32 fields
in the BLE ImuSnapshot proto. Only consumer is ble_service.c (already
assigning into int32) plus the driver bodies; rev7 build re-verified.

Verified on Rev 6 silicon over BLE: gravity tracks the correct axis at
~1000 mg, gyro responds to rotation. imu_init is now a real probe, so a
missing/dead IMU correctly trips the fatal park per the failure policy.

Builds: rev6 production, rev7 production. Flashed + live-confirmed.

+++ main/hal/imu_iface.h
@@ -30,14 +30,17 @@ extern "C" {
  */
 typedef struct {
     uint64_t timestamp_us;     ///< Capture time (board epoch, microseconds).
-    int16_t  accel_x_mg;       ///< milli-g (1 mg = 9.80665 mm/s^2).
-    int16_t  accel_y_mg;
-    int16_t  accel_z_mg;
-    int16_t  gyro_x_mdps;      ///< milli-degrees per second.
-    int16_t  gyro_y_mdps;
-    int16_t  gyro_z_mdps;
+    int32_t  accel_x_mg;       ///< milli-g (1 mg = 9.80665 mm/s^2).
+    int32_t  accel_y_mg;
+    int32_t  accel_z_mg;
+    int32_t  gyro_x_mdps;      ///< milli-degrees per second.
+    int32_t  gyro_y_mdps;
+    int32_t  gyro_z_mdps;
     uint16_t flags;            ///< IMU_SAMPLE_FLAG_* bitfield.
 } imu_sample_t;
+/* Fields are int32: at 500 dps full-scale a gyro reading reaches ~573 dps =
+ * 573000 mdps, far past int16's ±32767. Matches the SINT32 fields in the BLE
+ * ImuSnapshot proto so the value passes to the app without clamping. */
 
 #define IMU_SAMPLE_FLAG_NONE              0x0000
 #define IMU_SAMPLE_FLAG_FIFO_OVERFLOW     0x0001
